refactor: merge linklayerdevice domain with network
Linklayerdevice had a few types which have moved to the network domain.
The methods were not used.
