--- domain/linklayerdevice/doc.go
@@ -1,8 +1,0 @@
-// Copyright 2024 Canonical Ltd.
-// Licensed under the AGPLv3, see LICENCE file for details.
-
-package linklayerdevice
-
-// Package linklayerdevice deals with the enumerated types used with link layer devices.
-// It defines the type constants and has tests to ensure the values in the db lookup
-// tables match the enum values. It also has functions to convert between db and domain values.
