--- domain/linklayerdevice/package_test.go
@@ -1,14 +1,0 @@
-// Copyright 2024 Canonical Ltd.
-// Licensed under the AGPLv3, see LICENCE file for details.
-
-package linklayerdevice
-
-import (
-	"testing"
-
-	"github.com/juju/tc"
-)
-
-func TestPackage(t *testing.T) {
-	tc.TestingT(t)
-}
